requirements.txt
flask==3.0.3
requests==2.32.3
python-dateutil==2.9.0.post0
